+++ spec/spec_helper.rb
@@ -9,7 +9,7 @@ require "bundler/setup"
 begin
   require "simplecov"
   SimpleCov.start do
-    add_filter "/spec/"
+    skip "/spec/"
     enable_coverage :branch
   end
 rescue LoadError